# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Bank API",
        "contact": {
            "name": "API Support",
            "url": "http://www.mladenovic13.com",
            "email": "mladenovic13.dev@gmail.com"
        },
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/admin/request": {
            "get": {
                "security": [
                    {
                        "Bearer Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Get all requests",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/models.Request"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
        "/admin/request/:id/process": {
            "post": {
                "security": [
                    {
                        "Bearer Token": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Admin"
                ],
                "summary": "Process request",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Request ID",
                        "name": "requestID",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Request"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handler.Error"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "handler.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "models.Currency": {
            "type": "string",
            "enum": [
                "RSD",
                "USD",
                "USD"
            ],
            "x-enum-varnames": [
                "RSD",
                "USD",
                "EUR"
            ]
        },
        "models.Request": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "currency": {
                    "$ref": "#/definitions/models.Currency"
                },
                "id": {
                    "type": "string"
                },
                "isProcessed": {
                    "type": "boolean"
                },
                "type": {
                    "$ref": "#/definitions/models.RequestType"
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.RequestType": {
            "type": "string",
            "enum": [
                "OPEN_ACCOUNT",
                "CLOSE_ACCOUNT",
                "DEPOSIT",
                "WITHDRAW"
            ],
            "x-enum-varnames": [
                "OPEN_ACCOUNT",
                "CLOSE_ACCOUNT",
                "DEPOSIT",
                "WITHDRAW"
            ]
        },
...
    }
}